[ci] Extract test chunking helper
