+++ .changeset/extract-into-chunks.md
@@ -1,0 +1,8 @@
+---
+'@vercel/go': patch
+'@vercel/node': patch
+'@vercel/python': patch
+'@vercel/static-build': patch
+---
+
+Extract the shared integration-test chunking helper from the CI planner.
