+++ packages/static-build/test/integration-setup.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const { intoChunks } = require('../../../utils/chunk-tests');
+const { intoChunks } = require('../../../utils/into-chunks');
 
 const {
   testDeployment,
